moved addButtons components
