+++ src/components/addButton/AddButton.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
-import { Icon } from "./utils/Icon";
-import styles from "../styles/addButton.module.css";
+import { Icon } from "../utils/Icon";
+import styles from "../../styles/addButton.module.css";
 import { AddMenu } from "./AddMenu";
-import { List } from "../types/types";
+import { List } from "../../types/types";
 
 const AddButton = ({ activeList }: { activeList: List }) => {
   const [showAddMenu, setShowAddMenu] = useState(false);
